move query-direction to deprecated folders
