+++ packages/graphql/tests/tck/deprecated/query-direction-connection.test.ts
@@ -20,15 +20,60 @@
 import { Neo4jGraphQL } from "../../../src";
 import { formatCypher, formatParams, translateQuery } from "../utils/tck-test-utils";
 
-describe("QueryDirection in relationships connection", () => {
+describe("QueryDirection in relationships connection (deprecated _DEFAULT/_ONLY options)", () => {
     let typeDefs: string;
     let neoSchema: Neo4jGraphQL;
 
-    test("query connection with a DIRECTED relationship", async () => {
+    test("query with directed and undirected relationships with a DEFAULT_UNDIRECTED", async () => {
         typeDefs = /* GraphQL */ `
             type User @node {
                 name: String!
-                friends: [User!]! @relationship(type: "FRIENDS_WITH", direction: OUT, queryDirection: DIRECTED)
+                friends: [User!]!
+                    @relationship(type: "FRIENDS_WITH", direction: OUT, queryDirection: DEFAULT_UNDIRECTED)
+            }
+        `;
+
+        neoSchema = new Neo4jGraphQL({
+            typeDefs,
+        });
+        const query = /* GraphQL */ `
+            query FriendsAggregate {
+                users {
+                    friendsConnection {
+                        totalCount
+                    }
+                }
+            }
+        `;
+
+        const result = await translateQuery(neoSchema, query);
+
+        expect(formatCypher(result.cypher)).toMatchInlineSnapshot(`
+            "MATCH (this:User)
+            CALL {
+                WITH this
+                MATCH (this)-[this0:FRIENDS_WITH]-(this1:User)
+                WITH collect({ node: this1, relationship: this0 }) AS edges
+                WITH edges, size(edges) AS totalCount
+                CALL {
+                    WITH edges
+                    UNWIND edges AS edge
+                    WITH edge.node AS this1, edge.relationship AS this0
+                    RETURN collect({ node: { __id: id(this1), __resolveType: \\"User\\" } }) AS var2
+                }
+                RETURN { edges: var2, totalCount: totalCount } AS var3
+            }
+            RETURN this { friendsConnection: var3 } AS this"
+        `);
+
+        expect(formatParams(result.params)).toMatchInlineSnapshot(`"{}"`);
+    });
+
+    test("query connection with a DIRECTED_ONLY  relationship", async () => {
+        typeDefs = /* GraphQL */ `
+            type User @node {
+                name: String!
+                friends: [User!]! @relationship(type: "FRIENDS_WITH", direction: OUT, queryDirection: DIRECTED_ONLY)
             }
         `;
 
@@ -68,11 +113,11 @@ describe("QueryDirection in relationships connection", () => {
         expect(formatParams(result.params)).toMatchInlineSnapshot(`"{}"`);
     });
 
-    test("query with a UNDIRECTED relationship", async () => {
+    test("query with a UNDIRECTED_ONLY  relationship", async () => {
         typeDefs = /* GraphQL */ `
             type User @node {
                 name: String!
-                friends: [User!]! @relationship(type: "FRIENDS_WITH", direction: OUT, queryDirection: UNDIRECTED)
+                friends: [User!]! @relationship(type: "FRIENDS_WITH", direction: OUT, queryDirection: UNDIRECTED_ONLY)
             }
         `;
 
@@ -111,4 +156,6 @@ describe("QueryDirection in relationships connection", () => {
 
         expect(formatParams(result.params)).toMatchInlineSnapshot(`"{}"`);
     });
+
+   
 });
